requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
astroid==2.8.4
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coverage==6.1.1
Flask==2.0.2
flask-sock==0.4.0
Flask-WTF==0.15.1
fpdf==1.7.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pyaml==21.10.1
pycparser==2.20
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
simple-websocket==0.4.0
toml==0.10.2
tomli==1.2.2
Turbo-Flask==0.6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
WTForms==2.3.3
yarl==1.7.0
aiohttp>=3.7.4